changed task

+++ src/models/task.model.js
@@ -15,11 +15,11 @@ Task.init ({
         allowNull: false
     },
     latitude: {
-        type: DataTypes.DECIMAL,
+        type: DataTypes.FLOAT,
         allowNull: false
     },
     longitude: {
-        type: DataTypes.DECIMAL,
+        type: DataTypes.FLOAT,
         allowNull: false
     },
     hourStart: {
